Fix list assignment using extended slice with big integers

+++ Src/IronPython/Runtime/Slice.cs
@@ -101,7 +101,8 @@ namespace IronPython.Runtime {
 
         private static void DoSliceAssign(SliceAssign assign, int start, int stop, int step, object? value) {
             stop = step > 0 ? Math.Max(stop, start) : Math.Min(stop, start);
-            int n = Math.Max(0, (step > 0 ? (stop - start + step - 1) : (stop - start + step + 1)) / step);
+            // start, stop, or step may be near int.MaxValue so perform calculations in long
+            int n = (int)Math.Max(0, (step > 0 ? ((long)stop - start + step - 1) : ((long)stop - start + step + 1)) / step);
             // fast paths, if we know the size then we can
             // do this quickly.
             if (value is IList list) {
